+++ config/public/styles/global.css
@@ -15,13 +15,19 @@
   --color-9: #eeeeee;
   --color-10: #40ddb6;
   --color-11: #6b78e5;
+  --color-12: #098a69;
+  --color-13: #005540;
+  --color-14: #00795a;
+  --color-15: #6d6d6d;
 
   /* Font's */
   --h1: 700 1.62rem "Poppins", Arial, Sans-serif;
   --h2: 600 1.5rem "Poppins", Arial, Sans-serif;
   --h3: 500 1.3rem "Poppins", Arial, Sans-serif;
+  --h4: 500 1.15rem "Poppins", Arial, Sans-serif;
   --text-primary: 400 1rem "Poppins", Arial, Sans-serif;
   --text-secondary: 500 1rem "Poppins", Arial, Sans-serif;
+  --text-tertiary: 400 0.87rem "Poppins", Arial, Sans-serif;
   --text-detail: 400 0.87rem "Poppins", Arial, Sans-serif;
 
   /* Other's */
@@ -54,6 +60,10 @@ h3 {
   font: var(--h3);
 }
 
+h4 {
+  font: var(--h4);
+}
+
 /* ! Scroll Bar */
 /** The CSS code you provided is responsible for styling the scroll bar in a web page. */
 * {
@@ -84,7 +94,7 @@ h3 {
   justify-content: start;
 
   padding: 1.5rem 1rem;
-  gap: .65rem;
+  gap: 0.65rem;
 }
 
 .mt ._header img {
@@ -114,20 +124,21 @@ h3 {
   display: flex;
   align-items: center;
 
-  gap: .3rem;
+  gap: 0.3rem;
 }
 
 /* Main */
 .mt ._main {
-  display: grid;
-  place-content: center;
-
-  height: 100%;
+  display: flex;
+  align-items: flex-start;
+  justify-content: center;
+  gap: 8rem;
 
   padding: 3rem;
 }
 
-.mt ._data {
+.mt ._gerador,
+.mt ._preview {
   background: var(--color-3);
 
   width: 31.25rem;
@@ -138,17 +149,24 @@ h3 {
 
   gap: 1.5rem;
 
-  border-radius: .5rem;
+  border-radius: 0.5rem;
 }
 
-.mt ._data section {
+.mt ._gerador form,
+.mt ._preview ._cnt {
+  display: grid;
+
+  gap: 1.5rem;
+}
+
+.mt ._gerador section {
   display: flex;
   flex-direction: column;
 
   gap: 1rem;
 }
 
-.mt ._data .camp {
+.mt ._gerador .camp {
   display: flex;
   align-items: center;
 
@@ -156,25 +174,27 @@ h3 {
   color: var(--color-8);
 
   padding: 1rem;
-  gap: .8rem;
+  gap: 0.8rem;
 
   font: var(--text-primary);
   font-weight: 500;
 
-  border-radius: .5rem;
+  border-radius: 0.5rem;
+
+  cursor: text;
 }
 
-.mt ._data .textarea {
+.mt ._gerador .textarea {
   background: var(--color-4);
   color: var(--color-8);
 
-  border-radius: .5rem;
+  border-radius: 0.5rem;
 
   display: flex;
   align-items: start;
 
   padding: 1rem;
-  gap: .8rem;
+  gap: 0.8rem;
 
   font: var(--text-primary);
   font-weight: 500;
@@ -182,27 +202,207 @@ h3 {
   width: 100%;
 }
 
-.mt ._data .camp svg,
-.mt ._data .textarea svg {
+/* Select Styles */
+.mt ._gerador select {
+  appearance: none;
+  -webkit-appearance: none;
+  -moz-appearance: none;
+
+  text-indent: 1px;
+  text-overflow: "";
+
+  background-color: var(--color-4);
+
+  width: 100%;
+  height: 100%;
+}
+
+.chakra-select__wrapper.css-42b2qy {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.chakra-select__icon-wrapper.css-162mkon {
+  position: inherit;
+  transform: none;
+}
+
+.mt ._gerador .camp svg,
+.mt ._gerador .textarea svg,
+.mt ._gerador ._btn svg {
   font-size: 1.5rem;
 }
 
-.mt ._data .camp input,
-.mt ._data .textarea textarea,
-.mt ._data .camp select {
+.mt ._gerador .camp input,
+.mt ._gerador .textarea textarea {
   background: transparent;
 
   width: 100%;
   height: 100%;
 }
 
-.mt ._data .textarea textarea {
+.mt ._gerador .textarea textarea {
   height: 250px;
   resize: none;
 }
 
-.mt ._data .camp input::placeholder,
-.mt ._data .textarea textarea::placeholder {
+.mt ._gerador .camp input::placeholder,
+.mt ._gerador .textarea textarea::placeholder {
   color: var(--color-8);
 }
 
+.mt ._gerador ._btn {
+  display: flex;
+  align-items: center;
+
+  background: var(--color-13);
+  color: var(--color-8);
+
+  padding: 1rem;
+  gap: 0.8rem;
+
+  font: var(--text-primary);
+  font-weight: 500;
+
+  border-radius: 0.5rem;
+}
+
+.mt ._gerador ._btn:hover {
+  transition: all 0.5s;
+  background: var(--color-14);
+}
+
+/* Preview */
+.mt ._preview img {
+  border-radius: 10px;
+}
+
+.mt ._preview ._text {
+  display: flex;
+  flex-direction: column;
+
+  gap: 0.5rem;
+}
+
+.mt ._preview ._href {
+  font: var(--text-detail);
+  font-style: italic;
+
+  color: var(--color-12);
+
+  text-decoration: underline;
+  text-underline-offset: 3px;
+}
+
+.mt ._preview ._dt {
+  font: var(--text-tertiary);
+}
+
+/* Modal - Tag Gerada */
+.css-1oxhx2p {
+  width: 100vw !important;
+  height: 100vh !important;
+
+  background: rgba(34, 35, 39, 0.5);
+}
+
+.modal {
+  background: var(--color-5);
+  border-radius: 15px;
+
+  width: 65rem !important;
+
+  margin: auto;
+  padding: 35px 25px;
+  gap: 25px;
+
+  display: grid;
+  align-content: center;
+  justify-content: center;
+}
+
+.modal header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.modal header h3 {
+  font-size: 30px;
+}
+
+.modal header button {
+  padding: 0.5rem;
+
+  background-color: var(--color-3);
+  border-radius: 0.5rem;
+}
+
+.modal header button:hover {
+  transition: all 0.5s;
+
+  background-color: var(--color-14);
+}
+
+.modal header svg {
+  font-size: 1.5rem;
+}
+
+.modal textarea {
+  background: var(--color-3);
+
+  color: var(--color-6);
+
+  font: var(--text-primary);
+  font-weight: 500;
+
+  width: 100%;
+  height: 35rem;
+
+  padding: 1rem;
+  border-radius: 0.5rem;
+
+  resize: none;
+}
+
+.modal textarea::-webkit-scrollbar {
+  width: 12px;
+}
+
+.modal textarea::-webkit-scrollbar-track {
+  background: var(--color-3);
+  border-radius: 0.5rem;
+}
+
+.modal textarea::-webkit-scrollbar-thumb {
+  background-color: var(--color-15);
+  border-radius: 20px;
+
+  border: 4px solid var(--color-3);
+}
+
+.modal footer {
+  justify-content: center;
+}
+
+.modal ._btn {
+  display: flex;
+  align-items: center;
+
+  background: var(--color-13);
+  color: var(--color-8);
+
+  padding: 1rem;
+  gap: 0.8rem;
+
+  font: var(--text-primary);
+  font-weight: 500;
+
+  border-radius: 0.5rem;
+}
+
+.modal ._btn:hover {
+  transition: all 0.5s;
+  background: var(--color-14);
+}
